+++ app/service/print-all/page.tsx
@@ -1,0 +1,5 @@
+import PrintAllServiceSheets from "../../components/PrintAllServiceSheets";
+
+export default function PrintAllServiceSheetsPage() {
+  return <PrintAllServiceSheets />;
+}